Move is_container_installed and install_container to container module
