+++ dangerzone/util.py
@@ -1,6 +1,5 @@
 from __future__ import annotations
 
-import gzip
 import inspect
 import os
 import pathlib
@@ -13,8 +12,6 @@ import appdirs
 # If a general-purpose function or constant doesn't depend on anything else in the dangerzone package,
 # then it belongs here.
 from colorama import Back, Fore, Style
-
-import dangerzone
 
 SYSTEM = platform.system()
 
@@ -251,83 +248,6 @@ OCR_LANGUAGES = {
 }
 
 
-def is_container_installed():
-    """
-    See if the podman container is installed. Linux only.
-    """
-    # Get the image id
-    with open(get_resource_path("image-id.txt")) as f:
-        expected_image_id = f.read().strip()
-
-    # See if this image is already installed
-    installed = False
-    found_image_id = subprocess.check_output(
-        [
-            CONTAINER_RUNTIME,
-            "image",
-            "list",
-            "--format",
-            "{{.ID}}",
-            dangerzone.util.CONTAINER_NAME,
-        ],
-        text=True,
-        startupinfo=get_subprocess_startupinfo(),
-    )
-    found_image_id = found_image_id.strip()
-
-    if found_image_id == expected_image_id:
-        installed = True
-    elif found_image_id == "":
-        pass
-    else:
-        print("Deleting old dangerzone container image")
-
-        try:
-            subprocess.check_output(
-                [CONTAINER_RUNTIME, "rmi", "--force", found_image_id],
-                startupinfo=get_subprocess_startupinfo(),
-            )
-        except:
-            print("Couldn't delete old container image, so leaving it there")
-
-    return installed
-
-
-def install_container():
-    """
-    Make sure the podman container is installed. Linux only.
-    """
-    if is_container_installed():
-        return
-
-    # Load the container into podman
-    print("Installing Dangerzone container image...")
-
-    p = subprocess.Popen(
-        [CONTAINER_RUNTIME, "load"],
-        stdin=subprocess.PIPE,
-        startupinfo=get_subprocess_startupinfo(),
-    )
-
-    chunk_size = 10240
-    compressed_container_path = get_resource_path("container.tar.gz")
-    with gzip.open(compressed_container_path) as f:
-        while True:
-            chunk = f.read(chunk_size)
-            if len(chunk) > 0:
-                p.stdin.write(chunk)
-            else:
-                break
-    p.communicate()
-
-    if not is_container_installed():
-        print("Failed to install the container image")
-        return False
-
-    print("Container image installed")
-    return True
-
-
 def display_banner():
     """
     Raw ASCII art example:
@@ -483,7 +403,7 @@ def display_banner():
         + Back.BLACK
         + Fore.LIGHTWHITE_EX
         + Style.BRIGHT
-        + f"{' '*left_spaces}Dangerzone v{VERSION}{' '*right_spaces}"
+        + f"{' ' * left_spaces}Dangerzone v{VERSION}{' ' * right_spaces}"
         + Fore.YELLOW
         + Style.DIM
         + "│"
